backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
tavily-python==0.5.0
beautifulsoup4==4.12.3
httpx==0.27.2
pytesseract==0.3.13
Pillow==10.4.0
motor==3.5.1
pymongo==4.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.9
